requirements.txt
Flask
python-docx
PyMuPDF
transformers
torch
